+++ app/(auth)/lib/get-user-team-id.ts
