+++ app/(auth)/lib/index.ts
@@ -1,9 +1,9 @@
 // export { getUserInitializationTask } from "./get-initialization-task";
 export { deleteOauthCredential } from "./delete-oauth-credential";
 export { getAuthCallbackUrl } from "./get-auth-callback-url";
-export { getUserTeamId } from "./get-current-subscription";
 export { getCurrentTeam } from "./get-current-team";
 export { getOauthCredential } from "./get-oauth-credential";
+export { getUserTeamId } from "./get-user-team-id";
 export { isRoute06User } from "./is-route06-user";
 export { refreshOauthCredential } from "./refresh-oauth-credential";
 export type { Provider } from "./types";
